feat:#9-신고 제재 처리 로직 구현

+++ back/DevC/src/main/java/com/back/devc/domain/interaction/report/dto/AdminReportRequestDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AdminReportRequestDTO {
-    private Integer reportId;
+    private Long reportId;
     private String sanctionType;  // 제재 유형 (예: WARNING, TEMP_BAN, PERM_BAN)
     private String rejectReason;  // 반려 시 사유
 }
